Fix gumroad parsing error when normalizing data from an IO
Since the data may be coming in as an IO object, we need to be able to
read and rewind it and use the read value throughout the parsed object.

This is making me think it's about time to create an `ParsedEvent`
object that can keep that state in it; but I think that's a bigger
refactor than I have time for today.