Revise Roadmap for Q1~Q2 2020
Paired with Cheryl to figure out a roadmap for compensated; it will likely take a few dozen patches in order to complete them all; but we are pretty confident that these are the broad strokes that make sense for moving this to a revenue generating product.

We are probably going to pull some of the details into github issues; but may not have time today to do much of that.