Revise contributing with more useful information
Lane Grant recommended we put together a more concrete guide to getting started
contributing to Compensated. While this does not resolve the technical side of
that (getting CI, making sure bin/setup actually works, etc) it may serve as an
aspirational document that we can start filling in the missing gaps
around?