♻️ Let's make credit a top-level concern
I like the addition of the `CREDITS.md` file, and was
unsure if there was a convention between that and `CONTRIBUTORS.md`.

I looked here:
https://help.github.com/en/github/building-a-strong-community/creating-a-default-community-health-file

But didn't find one; so I'm going to leave it as CREDITS for now and put it at the top level.

I've also added @drllau's credit!