🚧 Begin to add support for canceled/ended
While working with @etlund, we discovered that we do not currently handle
when a stripe event comes in withended_at or cancelled_at values.

We began to restructure the subscription data structure to include the
subscription period and status.

However as we were testing, we wanted to add a permutation test for
when the canceled_at was set but ended_at wasn't, which prompted
a desire to interpolate data in our specs so we don't have to copy-paste
the fixtures over and over.

We think that using ERB for now makes sense, since it's built into Ruby.