✅➕Create persistent sandbox for running Compensated in feature tests
A Sandbox lets us isolate test environments for each scenario, allowing
for unlimited parallelization of feature-tests. Because compensated is
only a command-line program, the sandbox mostly contains the files,
but I could see it spinning up a tiny-ass web-server to be a fake
Stripe API kind of like MailCatcher does.

Still have to think about how to make the `Then` statements work.