🎉💡📝✅ Sprout cucumber-js for system-testing compensated (#80)
* 🎉💡📝✅ Sprout cucumber-js for system-testing compensated

See: https://github.com/zinc-collective/compensated/issues/79

I went back and forth between cucumber-js and cucumber-rb; and I wound
up choosing cucumber-js because JavaScript is the lingua-franca of the
web; and it seems like it makes it a bit easier for folks to get started
contributing to systems testing when we use a language that they're
familiar with.

This also adds the system tests to CI.

* ✨Sprout compensated executable

* ✅➕Create persistent sandbox for running Compensated in feature tests

A Sandbox lets us isolate test environments for each scenario, allowing
for unlimited parallelization of feature-tests. Because compensated is
only a command-line program, the sandbox mostly contains the files,
but I could see it spinning up a tiny-ass web-server to be a fake
Stripe API kind of like MailCatcher does.

Still have to think about how to make the `Then` statements work.