Increase likelihood of early detection of dependency conflicts (#88)
Including a Gemfile.lock in gems results in developers being forced to
conform to the particular version of Ruby or the particular gems that
other developers are using. For libraries, this means there's less
likelihood of developers running into compatibility issues on their
machines, which makes production releases a bit more brittle.

Ideally, we would wind up putting in some method for running our tests
across different combinations of dependencies, but for now uncommitting
the Gemfile.lock is a good step.