Wire in assertion that the appropriate products are created in Stripe
See: https://github.com/zinc-collective/compensated/issues/79

This is a big commit because it conflates the formatting changes from
https://github.com/zinc-collective/compensated/pull/89.

That said, it's also significant in it's own right. The ClientSandbox
now has the ability to query and assert against Stripe, so we will be
able to write full integration tests.

That said we _probably_ want to consider spinning out a fake Stripe API
so that our tests don't _always_ perform a full end-to-end check.

Perhaps in Coruru, if something like that doesn't exist already?