`compensated apply` creates products and prices in Stripe
See: https://github.com/zinc-collective/compensated/issues/79

The compensated cli now exists, and forwards options and configuration into
commands. Commands can be defined by extending the
`Compensated::Command` class, and overriding the `execute` method.

The first command is `apply`, which creates the configured products and
prices in the Payment Processors that support it.

This is a pretty-close to direct port of the code that I wrote to
allow folks to create products and prices in Stripe.

See: https://github.com/zinc-collective/convene/pull/26

Further, this gives us a `.rubocop.yml` file in the top level of the
project so that formatting can begin to become consistent across files.